% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_q1_mean_indices}
\alias{calculate_q1_mean_indices}
\title{Calculate Q1 and Mean Indices}
\usage{
calculate_q1_mean_indices(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named list containing the indices of the first quartile (Q1) and the mean.
}
\description{
This function calculates the indices of the first quartile (Q1) and the mean of a numeric vector.
}
\examples{
calculate_q1_mean_indices(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
