% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_youtube_data}
\alias{filter_youtube_data}
\title{Filter YouTube Data}
\usage{
filter_youtube_data(youtube_data, nb_subscribers, channel_types, countries)
}
\arguments{
\item{youtube_data}{A data frame containing YouTube channel statistics.}

\item{nb_subscribers}{A vector of length 2 specifying the minimum and maximum number of subscribers.}

\item{countries}{A vector of countries to filter.}

\item{categories}{A vector of categories to filter.}
}
\value{
A data frame containing the filtered YouTube channel statistics.
}
\description{
Filters the YouTube channel statistics data based on the given criteria.
}
\examples{
youtube_channel_stats <- data.frame(
  formatted_subscribers = c("1K", "10K", "100K", "1M"),
  channel_type = c("Music", "Games", "Entertainment", "Education"),
  country = c("USA", "UK", "USA", "UK")
)
filtered_data <- filter_youtube_data(
  youtube_data = youtube_channel_stats,
  nb_subscribers = c("1K", "1M"),
  channel_types = c("Education", "Entertainment"),
  countries = c("USA", "UK")
)
}
