% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{youtube_channel_stats}
\alias{youtube_channel_stats}
\title{YouTube Channel Statistics}
\format{
A data frame with statistics of YouTube channels.
}
\usage{
youtube_channel_stats
}
\description{
This dataset contains statistics of YouTube channels sourced from Kaggle.
}
\details{
The data includes various metrics such as rank, subscribers, video views,
category, country, earnings, and more for the top YouTube channels.

Source: \href{https://www.kaggle.com/datasets/nelgiriyewithana/global-youtube-statistics-2023/data}{Kaggle - Global YouTube Statistics 2023}
}
\keyword{datasets}
